refactor: move max_diff
And add tests
